Create a Flux with interval
Create a Flux that emits long values starting with 0 and incrementing at specified time intervals on the global timer. 
Why there are Thread.sleep? This method publishes elements at regular intervals while running on another thread. It doesn't block the main thread. So when the test class finishes, all the threads are killed and nothing happens. So what we can do is sleep the main threat for, let's say, 2 seconds to give time to the other thread to publish some values.
Notice that an onCcomplete call was not logged. This Flux will never complete. The only reason it stopped was because the thread was killed when the program finished.
in a web server, for example, that is always running, we could have a publisher generating an infinite stream of values at a regular interval. In cases like this one, we could use the take operator to take only the first N values from the Flux.